Move functions to lib.py
